+++ dags/common/overrides_manager.py
@@ -1,13 +1,15 @@
+import time
 import uuid
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
 
 # Forward declaration for type hints - actual import at bottom to avoid circular dependency
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Generic, TypeVar
 
 from clickhouse_driver import Client
 
 from posthog import settings
+from posthog.clickhouse.cluster import AlterTableMutationRunner, LightweightDeleteMutationRunner
 
 if TYPE_CHECKING:
     pass
@@ -55,3 +57,101 @@ class OverridesSnapshotTable(ABC):
             {"database": settings.CLICKHOUSE_DATABASE, "table": self.name},
         )
         assert queue_size == 0
+
+
+TOverridesSnapshotTable = TypeVar("TOverridesSnapshotTable", bound=OverridesSnapshotTable)
+
+
+@dataclass
+class OverridesSnapshotDictionary(ABC, Generic[TOverridesSnapshotTable]):
+    source: TOverridesSnapshotTable
+
+    @property
+    def name(self) -> str:
+        return f"{self.source.name}_dictionary"
+
+    @property
+    def qualified_name(self):
+        return f"{settings.CLICKHOUSE_DATABASE}.{self.name}"
+
+    def create(self, client: Client, shards: int, max_execution_time: int, max_memory_usage: int) -> None:
+        raise NotImplementedError()
+
+    def exists(self, client: Client) -> bool:
+        results = client.execute(
+            "SELECT count() FROM system.dictionaries WHERE database = %(database)s AND name = %(name)s",
+            {"database": settings.CLICKHOUSE_DATABASE, "name": self.name},
+        )
+        [[count]] = results
+        return count > 0
+
+    def drop(self, client: Client) -> None:
+        client.execute(f"DROP DICTIONARY IF EXISTS {self.qualified_name} SYNC")
+
+    def __is_loaded(self, client: Client) -> bool:
+        results = client.execute(
+            "SELECT status, last_exception FROM system.dictionaries WHERE database = %(database)s AND name = %(name)s",
+            {"database": settings.CLICKHOUSE_DATABASE, "name": self.name},
+        )
+        if not results:
+            raise Exception("dictionary does not exist")
+        else:
+            [[status, last_exception]] = results
+            if status == "LOADED":
+                return True
+            elif status in {"LOADING", "FAILED_AND_RELOADING", "LOADED_AND_RELOADING"}:
+                return False
+            elif status == "FAILED":
+                raise Exception(f"failed to load: {last_exception}")
+            else:
+                raise Exception(f"unexpected status: {status}")
+
+    def load(self, client: Client):
+        # TODO: this should probably not reload if the dictionary is already loaded
+        client.execute(f"SYSTEM RELOAD DICTIONARY {self.qualified_name}")
+
+        # reload is async, so we need to wait for the dictionary to actually be loaded
+        # TODO: this should probably throw on unexpected reloads
+        while not self.__is_loaded(client):
+            time.sleep(5.0)
+
+        results = client.execute(
+            f"""
+            SELECT groupBitXor(row_checksum) AS table_checksum
+            FROM (SELECT cityHash64(*) AS row_checksum FROM {self.qualified_name} ORDER BY team_id, distinct_id)
+            """
+        )
+        [[checksum]] = results
+        return checksum
+
+    @property
+    def update_table(self):
+        raise NotImplementedError()
+
+    @property
+    def update_commands(self):
+        raise NotImplementedError()
+
+    @property
+    def update_mutation_runner(self) -> AlterTableMutationRunner:
+        return AlterTableMutationRunner(
+            table=self.update_table,
+            commands=self.update_commands,
+            parameters={"name": self.qualified_name},
+        )
+
+    @property
+    def overrides_table(self):
+        raise NotImplementedError()
+
+    @property
+    def overrides_deletes_predicate(self):
+        raise NotImplementedError()
+
+    @property
+    def overrides_delete_mutation_runner(self) -> LightweightDeleteMutationRunner:
+        return LightweightDeleteMutationRunner(
+            table=self.overrides_table,
+            predicate=self.overrides_deletes_predicate,
+            parameters={"name": self.qualified_name},
+        )
