+++ dags/person_overrides.py
@@ -1,4 +1,3 @@
-import time
 import uuid
 import datetime
 from dataclasses import dataclass
@@ -9,17 +8,12 @@ import pydantic
 from clickhouse_driver import Client
 
 from posthog import settings
-from posthog.clickhouse.cluster import (
-    AlterTableMutationRunner,
-    ClickhouseCluster,
-    LightweightDeleteMutationRunner,
-    MutationWaiter,
-)
+from posthog.clickhouse.cluster import ClickhouseCluster, MutationWaiter
 from posthog.models.event.sql import EVENTS_DATA_TABLE
 from posthog.models.person.sql import PERSON_DISTINCT_ID_OVERRIDES_TABLE
 
 from dags.common import JobOwners
-from dags.common.overrides_manager import OverridesSnapshotTable
+from dags.common.overrides_manager import OverridesSnapshotDictionary, OverridesSnapshotTable
 
 
 @dataclass
@@ -65,16 +59,8 @@ class PersonOverridesSnapshotTable(OverridesSnapshotTable):
 
 
 @dataclass
-class PersonOverridesSnapshotDictionary:
+class PersonOverridesSnapshotDictionary(OverridesSnapshotDictionary):
     source: PersonOverridesSnapshotTable
-
-    @property
-    def name(self) -> str:
-        return f"{self.source.name}_dictionary"
-
-    @property
-    def qualified_name(self):
-        return f"{settings.CLICKHOUSE_DATABASE}.{self.name}"
 
     def create(self, client: Client, shards: int, max_execution_time: int, max_memory_usage: int) -> None:
         client.execute(
@@ -99,70 +85,23 @@ class PersonOverridesSnapshotDictionary:
             },
         )
 
-    def exists(self, client: Client) -> bool:
-        results = client.execute(
-            "SELECT count() FROM system.dictionaries WHERE database = %(database)s AND name = %(name)s",
-            {"database": settings.CLICKHOUSE_DATABASE, "name": self.name},
-        )
-        [[count]] = results
-        return count > 0
-
-    def drop(self, client: Client) -> None:
-        client.execute(f"DROP DICTIONARY IF EXISTS {self.qualified_name} SYNC")
-
-    def __is_loaded(self, client: Client) -> bool:
-        results = client.execute(
-            "SELECT status, last_exception FROM system.dictionaries WHERE database = %(database)s AND name = %(name)s",
-            {"database": settings.CLICKHOUSE_DATABASE, "name": self.name},
-        )
-        if not results:
-            raise Exception("dictionary does not exist")
-        else:
-            [[status, last_exception]] = results
-            if status == "LOADED":
-                return True
-            elif status in {"LOADING", "FAILED_AND_RELOADING", "LOADED_AND_RELOADING"}:
-                return False
-            elif status == "FAILED":
-                raise Exception(f"failed to load: {last_exception}")
-            else:
-                raise Exception(f"unexpected status: {status}")
-
-    def load(self, client: Client):
-        # TODO: this should probably not reload if the dictionary is already loaded
-        client.execute(f"SYSTEM RELOAD DICTIONARY {self.qualified_name}")
-
-        # reload is async, so we need to wait for the dictionary to actually be loaded
-        # TODO: this should probably throw on unexpected reloads
-        while not self.__is_loaded(client):
-            time.sleep(5.0)
-
-        results = client.execute(
-            f"""
-            SELECT groupBitXor(row_checksum) AS table_checksum
-            FROM (SELECT cityHash64(*) AS row_checksum FROM {self.qualified_name} ORDER BY team_id, distinct_id)
-            """
-        )
-        [[checksum]] = results
-        return checksum
+    @property
+    def update_table(self):
+        return EVENTS_DATA_TABLE()
 
     @property
-    def person_id_update_mutation_runner(self) -> AlterTableMutationRunner:
-        return AlterTableMutationRunner(
-            table=EVENTS_DATA_TABLE(),
-            commands={
-                "UPDATE person_id = dictGet(%(name)s, 'person_id', (team_id, distinct_id)) WHERE dictHas(%(name)s, (team_id, distinct_id))"
-            },
-            parameters={"name": self.qualified_name},
-        )
+    def update_commands(self):
+        return {
+            "UPDATE person_id = dictGet(%(name)s, 'person_id', (team_id, distinct_id)) WHERE dictHas(%(name)s, (team_id, distinct_id))"
+        }
 
     @property
-    def overrides_delete_mutation_runner(self) -> LightweightDeleteMutationRunner:
-        return LightweightDeleteMutationRunner(
-            table=PERSON_DISTINCT_ID_OVERRIDES_TABLE,
-            predicate="isNotNull(dictGetOrNull(%(name)s, 'version', (team_id, distinct_id)) as snapshot_version) AND snapshot_version >= version",
-            parameters={"name": self.qualified_name},
-        )
+    def overrides_table(self):
+        return PERSON_DISTINCT_ID_OVERRIDES_TABLE
+
+    @property
+    def overrides_deletes_predicate(self):
+        return "isNotNull(dictGetOrNull(%(name)s, 'version', (team_id, distinct_id)) as snapshot_version) AND snapshot_version >= version"
 
 
 # Snapshot Table Management
@@ -297,7 +236,7 @@ def run_person_id_update_mutations(
     cluster: dagster.ResourceParam[ClickhouseCluster],
     dictionary: PersonOverridesSnapshotDictionary,
 ) -> PersonOverridesSnapshotDictionary:
-    dictionary.person_id_update_mutation_runner.run_on_shards(cluster)
+    dictionary.update_mutation_runner.run_on_shards(cluster)
     return dictionary
